// src/generated/models/Lm_leavetypesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Lm_leavetypeslm_isactive = {
  0: 'No',
  1: 'Yes'
} as const;
export type Lm_leavetypeslm_isactive = keyof typeof Lm_leavetypeslm_isactive;
export const Lm_leavetypeslm_requiresapproval = {
  0: 'No',
  1: 'Yes'
} as const;
export type Lm_leavetypeslm_requiresapproval = keyof typeof Lm_leavetypeslm_requiresapproval;
export const Lm_leavetypesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Lm_leavetypesstatecode = keyof typeof Lm_leavetypesstatecode;
export const Lm_leavetypesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Lm_leavetypesstatuscode = keyof typeof Lm_leavetypesstatuscode;

export interface Lm_leavetypesBase {
  importsequencenumber?: string;
  lm_code: string;
  lm_color?: string;
  lm_icon?: string;
  lm_isactive?: Lm_leavetypeslm_isactive;
  lm_leavetypeid: string;
  lm_maxdaysperrequest?: string;
  lm_name: string;
  lm_requiresapproval?: Lm_leavetypeslm_requiresapproval;
  lm_sortorder?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Lm_leavetypesstatecode;
  statuscode?: Lm_leavetypesstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Lm_leavetypes extends Lm_leavetypesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  lm_isactivename?: string;
  lm_requiresapprovalname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}