// src/generated/models/Lm_publicholidaiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Lm_publicholidaieslm_isrecurring = {
  0: 'No',
  1: 'Yes'
} as const;
export type Lm_publicholidaieslm_isrecurring = keyof typeof Lm_publicholidaieslm_isrecurring;
export const Lm_publicholidaiesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Lm_publicholidaiesstatecode = keyof typeof Lm_publicholidaiesstatecode;
export const Lm_publicholidaiesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Lm_publicholidaiesstatuscode = keyof typeof Lm_publicholidaiesstatuscode;

export interface Lm_publicholidaiesBase {
  importsequencenumber?: string;
  lm_date?: string;
  lm_isrecurring?: Lm_publicholidaieslm_isrecurring;
  lm_name: string;
  lm_publicholidayid: string;
  lm_region?: string;
  overriddencreatedon?: string;
  statecode: Lm_publicholidaiesstatecode;
  statuscode?: Lm_publicholidaiesstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Lm_publicholidaies extends Lm_publicholidaiesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  lm_isrecurringname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}